Register uncaught exception handler for Node thread
Handler calls #destroy on the Node's ZContext, avoiding the case in
which an unexpected exception occurs and leaves the process hanging
indefinitely because the context is otherwise never shut down.

This is actually hard to test for. Currently, the change made to the
test in FloodFillTests will cause the test to fail, but without this
uncaught exception handler, the test would never finish at all.